+++ lib/trailblazer/activity/dsl/linear.rb
@@ -179,6 +179,7 @@ end
 
 require "trailblazer/activity/dsl/linear/helper"
 require "trailblazer/activity/dsl/linear/normalizer"
+require "trailblazer/activity/dsl/linear/normalizer/terminus"
 require "trailblazer/activity/dsl/linear/helper/path"
 require "trailblazer/activity/dsl/linear/state"
 require "trailblazer/activity/dsl/linear/strategy"