+++ lib/trailblazer/activity/dsl/linear.rb
@@ -1,4 +1,5 @@
 require "trailblazer-activity"
+require "trailblazer/declarative"
 
 class Trailblazer::Activity
   module DSL