+++ lib/trailblazer/activity/dsl/linear/state.rb
@@ -10,30 +10,61 @@ module Trailblazer
         #
         # DISCUSS: why do we have this structure? It doesn't cover "immutable copying", that has to be done by its clients.
         #          also, copy with to_h
-        class State
+        class State # TODO: rename to Dsl
+          # @state = Declarative.State(tuples)
+# +            initialize_state!(
+# +              # "artifact/sequence" =>       [, {copy: Trailblazer::Declarative::State.method(:subclass)}],
+# +              # "dsl/recorded_options" => [Hash.new, {}], # copy # FIXME: we need real definitions here, I guess.
+# +            )
+          def self.build(normalizers:, initial_sequence:, fields: {}, **normalizer_options)
+            tuples = {
+              "sequence" =>       [initial_sequence, {}],
+              "dsl/normalizer" =>          [normalizers, {}],  # copy on inherit
+              "dsl/normalizer_options" => [normalizer_options, {}], # copy on inherit
+              "fields" => [fields, {}],
+            }
+
+            state = Trailblazer::Declarative.State(tuples)
+
+            new(state)
+          end
+
             # remembers how to call normalizers (e.g. track_color), TaskBuilder
             # remembers sequence
-          def initialize(normalizers:, initial_sequence:, fields: {}.freeze, **normalizer_options)
-            @normalizer         = normalizers # compiled normalizers.
-            @sequence           = initial_sequence
-            @normalizer_options = normalizer_options
-            @fields             = fields
+          def initialize(state)
+            @state = state
           end
 
           # Called to "inherit" a state.
           def copy
-            self.class.new(normalizers: @normalizer, initial_sequence: @sequence, fields: @fields, **@normalizer_options)
+            inherited_fields = @state.copy_fields()
+
+            state = Trailblazer::Declarative.State(inherited_fields) # FIXME: identical to State::initialize_state!
+
+            self.class.new(state)
+          end
+
+          def sequence
+            @state.get("sequence")
           end
 
           def to_h
+            raise
             {sequence: @sequence, normalizers: @normalizer, normalizer_options: @normalizer_options, fields: @fields} # FIXME.
           end
 
           def update_sequence(&block)
-            @sequence = yield(**to_h)
+            @state.update!("sequence") do |sequence|
+              yield(
+                sequence: sequence,
+                normalizers: @state.get("dsl/normalizer"),
+                normalizer_options: @state.get("dsl/normalizer_options") # FIXME: could we store this with the normalizers?
+              ) # FIXME: define interface for block.
+            end
           end
 
           def update_options(fields)
+            raise
             @fields = fields
           end
 
@@ -43,8 +74,8 @@ module Trailblazer
             #
             # We simply compile the activities that represent the normalizers for #step, #pass, etc.
             # This can happen at compile-time, as normalizers are stateless.
-            def initialize(normalizer_sequences)
-              @normalizers = normalizer_sequences
+            def initialize(normalizer_pipelines)
+              @normalizers = normalizer_pipelines
             end
 
             # Execute the specific normalizer (step, fail, pass) for a particular option set provided